+++ src/styles/scss/_header.scss
@@ -41,11 +41,9 @@
   flex-flow: row wrap;
   gap: 20px;
   justify-content: space-around;
-
   &__cart {
     position: relative;
   }
-
   &__link > svg {
     color: $color-yellow;
     width: 60px;
@@ -56,7 +54,6 @@
     transition-property: color, outline-color;
     transition-duration: 0.2s;
     transition-timing-function: ease-in-out;
-
     &:hover {
       color: $color-white-tr;
     }
@@ -67,7 +64,6 @@
       color: $color-grey-s-tr;
     }
   }
-
   &__amount {
     position: absolute;
     width: 20px;
