+++ src/styles/scss/_main_products.scss
@@ -5,10 +5,8 @@
     width: 73%;
   }
 }
-
 .main__products {
   padding: 6px;
-
   .products {
     padding-top: 6px;
     display: flex;
@@ -27,18 +25,12 @@
       align-self: stretch;
       justify-content: flex-end;
     }
-
-    // .products__title
-
     &__title {
       margin: 4px;
       color: $color-white;
       text-align: center;
       flex-grow: 1;
     }
-
-    // .products__photo
-
     &__photo {
       height: 300px;
       background-position: center;
@@ -46,17 +38,11 @@
       width: 98%;
       flex-grow: 1;
     }
-
-    // .products__price
-
     &__price {
       margin-top: 8px;
       color: $color-yellow;
       text-align: center;
     }
-
-    // .products__options
-
     &__options {
       display: flex;
       flex-flow: row wrap;
@@ -86,7 +72,6 @@
       }
     }
   }
-
   .products__top-panel {
     display: flex;
     justify-content: space-between;
