+++ src/styles/scss/_typography.scss
@@ -3,48 +3,40 @@ html {
   box-sizing: border-box;
   height: 100%;
 }
-
 body {
   background-color: $color-brown-xl;
   min-height: 100%;
   display: flex;
   flex-direction: column;
 }
-
 *,
 *::before,
 *::after {
   box-sizing: inherit;
 }
-
 * {
   margin: 0;
   padding: 0;
-  font-family: "Trebuchet MS", "Lucida Sans Unicode", "Lucida Grande",
-    "Lucida Sans", Arial, sans-serif;
+  font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial,
+    sans-serif;
 }
-
 a {
   text-decoration: none;
   outline: none;
 }
-
 svg {
   max-width: 100%;
   height: auto;
 }
-
 ul {
   list-style: none;
 }
-
 .section {
   display: flex;
   //flex-flow: column wrap;
   align-content: center;
   align-items: center;
 }
-
 button {
   border: none;
   font: inherit;
