+++ src/styles/scss/_variables.scss
@@ -16,7 +16,6 @@ $color-white-tr: rgba(255, 255, 255, 0.47);
   font-weight: normal;
   font-style: normal;
 }
-
 @font-face {
   font-family: 'Molgan';
   src: url('../assets/fonts/Molgan-Regular.otf') format('opentype');
