+++ src/styles/styles.scss
@@ -12,5 +12,4 @@
 @import '../pages/product-details/details';
 @import '../pages/purchase-page/purchase';
 @import './scss/media';
-
 @import './scss/active';
